+++ packages/website/src/pages/advanced/SlideY.tsx
@@ -1,3 +1,4 @@
+import {SlideYMetadata} from '@coveord/plasma-components-props-analyzer';
 import code from '@examples/SlideY/main.example.tsx';
 
 import {PageLayout} from '../../building-blocs/PageLayout';
@@ -10,5 +11,6 @@ export default () => (
         description="Allows to hide and show content using a vertical expand animation."
         sourcePath="packages/react/src/animations/SlideY.tsx"
         code={code}
+        propsMetadata={SlideYMetadata}
     />
 );