+++ packages/website/src/pages/form/MultilineBox.tsx
@@ -1,8 +1,9 @@
+import {MultilineBoxMetadata} from '@coveord/plasma-components-props-analyzer';
 import code from '@examples/multiline-box/MultilineBox.example.tsx';
-import withContainer from '@examples/multiline-box/MultilineBoxWithContainer.example.tsx';
-import withRemove from '@examples/multiline-box/MultilineBoxWithRemove.example.tsx';
-import withDragAndDrop from '@examples/multiline-box/MultilineBoxWithDragAndDrop.example.tsx';
 import complex from '@examples/multiline-box/MultilineBoxComplex.example.tsx';
+import withContainer from '@examples/multiline-box/MultilineBoxWithContainer.example.tsx';
+import withDragAndDrop from '@examples/multiline-box/MultilineBoxWithDragAndDrop.example.tsx';
+import withRemove from '@examples/multiline-box/MultilineBoxWithRemove.example.tsx';
 
 import {PageLayout} from '../../building-blocs/PageLayout';
 
@@ -13,6 +14,7 @@ export default () => (
         description="A multiline box allows users to provide multiple inputs for the same parameter. Each input appears on a different line."
         section="Form"
         code={code}
+        propsMetadata={MultilineBoxMetadata}
         examples={{
             withContainer: {code: withContainer, title: 'Custom container'},
             withRemove: {code: withRemove, title: 'Remove button'},