+++ packages/website/src/pages/form/Slider.tsx
@@ -1,7 +1,8 @@
+import {SliderMetadata} from '@coveord/plasma-components-props-analyzer';
 import code from '@examples/Slider/Slider.example.tsx';
+import append from '@examples/Slider/SliderAppend.example.tsx';
 import asymetric from '@examples/Slider/SliderAsymetric.example.tsx';
 import onChange from '@examples/Slider/SliderOnChange.example.tsx';
-import append from '@examples/Slider/SliderAppend.example.tsx';
 
 import {PageLayout} from '../../building-blocs/PageLayout';
 
@@ -13,6 +14,7 @@ export default () => (
         description="A slider offers a quick and visual way for users to select a value within a given range."
         componentSourcePath="/slider/Slider.tsx"
         code={code}
+        propsMetadata={SliderMetadata}
         examples={{
             asymetric: {code: asymetric, title: 'Asymetric'},
             onChange: {code: onChange, title: 'Change handler'},