+++ packages/website/src/pages/layout/Table.tsx
@@ -1,13 +1,14 @@
+import {TableHOCMetadata} from '@coveord/plasma-components-props-analyzer';
 import code from '@examples/Table/main.example.tsx';
-import withLoading from '@examples/Table/withLoading.example.tsx';
+import withAction from '@examples/Table/withAction.example.tsx';
 import withBlankslate from '@examples/Table/withBlankslate.example.tsx';
+import withDatePicker from '@examples/Table/withDatePicker.example.tsx';
+import withFilter from '@examples/Table/withFilter.example.tsx';
+import withLoading from '@examples/Table/withLoading.example.tsx';
 import withPagination from '@examples/Table/withPagination.example.tsx';
 import withPredicate from '@examples/Table/withPredicate.example.tsx';
-import withAction from '@examples/Table/withAction.example.tsx';
-import withSort from '@examples/Table/withSort.example.tsx';
-import withFilter from '@examples/Table/withFilter.example.tsx';
-import withDatePicker from '@examples/Table/withDatePicker.example.tsx';
 import withServer from '@examples/Table/withServer.example.tsx';
+import withSort from '@examples/Table/withSort.example.tsx';
 
 import {PageLayout} from '../../building-blocs/PageLayout';
 
@@ -19,6 +20,7 @@ const TableExamples = () => (
         section="Layout"
         description="A table displays large quantities of items or data in a list format. Filtering features and actions may be added."
         code={code}
+        propsMetadata={TableHOCMetadata}
         examples={{
             withLoading: {code: withLoading, title: 'Loading table'},
             withBlankslate: {code: withBlankslate, title: 'Table blankslate'},