+++ packages/website/src/pages/navigation/SideNavigation.tsx
@@ -1,3 +1,5 @@
+import {SideNavigationMetadata} from '@coveord/plasma-components-props-analyzer';
+
 import {PageLayout} from '../../building-blocs/PageLayout';
 import collapsible from '../../examples/SideNavigation/collapsible.example.tsx';
 import loading from '../../examples/SideNavigation/loading.example.tsx';
@@ -12,6 +14,7 @@ export const SideNavigationExamples = () => (
         description="A sidebar navigation is a primary navigation element that displays the architecture of the product’s features."
         thumbnail="sideNav"
         code={code}
+        propsMetadata={SideNavigationMetadata}
         examples={{
             loading: {code: loading, title: 'Loading section'},
             collapsible: {code: collapsible, title: 'Collapsible section'},